+++ crates/ika-move-packages/packages/ika_system/sources/dwallet_2pc_mpc_coordinator.move
@@ -110,7 +110,7 @@ public fun get_active_encryption_key(
 
 public fun register_encryption_key(
     self: &mut DWalletCoordinator,
-    curve: u8,
+    curve: u32,
     encryption_key: vector<u8>,
     encryption_key_signature: vector<u8>,
     signer_public_key: vector<u8>,
@@ -128,8 +128,8 @@ public fun register_encryption_key(
 public fun approve_message(
     self: &mut DWalletCoordinator,
     dwallet_cap: &DWalletCap,
-    signature_algorithm: u8,
-    hash_scheme: u8,
+    signature_algorithm: u32,
+    hash_scheme: u32,
     message: vector<u8>
 ): MessageApproval {
     self.inner().approve_message(
@@ -143,8 +143,8 @@ public fun approve_message(
 public fun approve_imported_key_message(
     self: &mut DWalletCoordinator,
     imported_key_dwallet_cap: &ImportedKeyDWalletCap,
-    signature_algorithm: u8,
-    hash_scheme: u8,
+    signature_algorithm: u32,
+    hash_scheme: u32,
     message: vector<u8>
 ): ImportedKeyMessageApproval {
     self.inner().approve_imported_key_message(
@@ -158,7 +158,7 @@ public fun approve_imported_key_message(
 public fun request_dwallet_dkg_first_round(
     self: &mut DWalletCoordinator,
     dwallet_network_decryption_key_id: ID,
-    curve: u8,
+    curve: u32,
     payment_ika: &mut Coin<IKA>,
     payment_sui: &mut Coin<SUI>,
     ctx: &mut TxContext
@@ -200,7 +200,7 @@ public fun request_dwallet_dkg_second_round(
 public fun new_imported_key_dwallet(
     self: &mut DWalletCoordinator,
     dwallet_network_decryption_key_id: ID,
-    curve: u8,
+    curve: u32,
     ctx: &mut TxContext
 ): ImportedKeyDWalletCap {
     self.inner_mut().new_imported_key_dwallet(
@@ -272,7 +272,7 @@ public fun accept_encrypted_user_share(
 public fun request_presign(
     self: &mut DWalletCoordinator,
     dwallet_id: ID,
-    signature_algorithm: u8,
+    signature_algorithm: u32,
     payment_ika: &mut Coin<IKA>,
     payment_sui: &mut Coin<SUI>,
     ctx: &mut TxContext
@@ -289,8 +289,8 @@ public fun request_presign(
 public fun request_global_presign(
     self: &mut DWalletCoordinator,
     dwallet_network_decryption_key_id: ID,
-    curve: u8,
-    signature_algorithm: u8,
+    curve: u32,
+    signature_algorithm: u32,
     payment_ika: &mut Coin<IKA>,
     payment_sui: &mut Coin<SUI>,
     ctx: &mut TxContext
@@ -357,7 +357,7 @@ public fun request_future_sign(
     dwallet_id: ID,
     presign_cap: PresignCap,
     message: vector<u8>,
-    hash_scheme: u8,
+    hash_scheme: u32,
     message_centralized_signature: vector<u8>,
     payment_ika: &mut Coin<IKA>,
     payment_sui: &mut Coin<SUI>,
